making api cleaner for entry point deciphering
